+++ DEVinBank/Extrato.cs
@@ -10,12 +10,12 @@ namespace DEVinBank
     {
        public Conta? ContaDestino { get; private set; }
         public Conta ContaOrigem { get; private set; }
-        public decimal Valor { get; private set; }
+        public double Valor { get; private set; }
       
         public string Tipo { get; private set; }
         public DateTime Date { get; private set; }
 
-        public Extrato(decimal valor, Conta contaOrigem, Conta contaDestino, string tipo)
+        public Extrato(double valor, Conta contaOrigem, Conta contaDestino, string tipo)
         {
             ContaOrigem = contaOrigem;
             ContaDestino = contaDestino;
@@ -23,7 +23,7 @@ namespace DEVinBank
             Tipo = tipo;
             Date = DateTime.Now;
         }
-        public Extrato(decimal valor, Conta contaOrigem, string tipo)
+        public Extrato(double valor, Conta contaOrigem, string tipo)
         {
             ContaOrigem = contaOrigem;
             Valor = valor;
